requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
pyyaml>=6.0
matplotlib>=3.7.0
plotly>=5.15.0
scipy>=1.10.0
pytest>=7.0.0
ta-lib>=0.4.0
rich>=13.0.0
python-dotenv>=1.0.0
scikit-learn>=1.2.0
lightgbm>=3.3.0
schedule>=1.2.0
joblib>=1.2.0
